文章管理三个接口第一次提交

+++ src/main/java/com/huotu/ymr/model/AppArticleListModel.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AppArticleListModel {
-    private Integer pId;
+    private Long pId;
 
 
     /**
